requirements.txt
mcp[cli]
httpx
google-genai
python-dotenv
